requirements.txt
streamlit>=1.24.0
opencv-python>=4.7.0
numpy>=1.21.0
pillow>=9.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
plotly>=5.3.0
scikit-learn>=0.24.2
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
supervision>=0.6.0